requirements.txt
Django==3.0.6
djangorestframework==3.11.0
pyjwt==1.0.0
bcrypt==3.1.7